+++ webclient/src/components/Common/useDebouncedCallback.js
@@ -3,6 +3,8 @@
 import { useCallback, useEffect, useRef } from "react";
 
 export default function useDebouncedCallback(callback, delay, deps) {
+  const reject = useRef(null);
+  const resolve = useRef(null);
   const functionTimeoutHandler = useRef(null);
   const debouncedFunction = useCallback(callback, deps);
 
@@ -16,7 +18,19 @@ export default function useDebouncedCallback(callback, delay, deps) {
   return (...args) => {
     clearTimeout(functionTimeoutHandler.current);
     functionTimeoutHandler.current = setTimeout(() => {
-      debouncedFunction(...args);
+      try {
+        const output = debouncedFunction(...args);
+        resolve.current(output);
+      } catch (e) {
+        reject.current(e);
+      }
     }, delay);
+
+    const promise = new Promise((res, rej) => {
+      resolve.current = res;
+      reject.current = rej;
+    });
+
+    return promise;
   };
 }