+++ webclient/src/components/product/Form/actionTypes.js
@@ -1,0 +1,12 @@
+export const NAME_CHANGE = 'NAME_FIELD_CHANGE';
+export const CATEGORY_CHANGE = 'CATEGORY_FIELD_CHANGE';
+export const EXPIRY_CHANGE = 'EXPIRY_CHANGE';
+export const IS_EXPIRY_CHANGE = 'IS_EXPIRY_CHANGE';
+export const PRICE_CHANGE = 'PRICE_CHANGE';
+export const SIZE_CHANGE = 'SIZE_CHANGE';
+export const DESCRIPTION_CHANGE = 'DESCRIPTION_CHANGE';
+
+export const SAVE = 'SAVE';
+export const SAVE_INIT = 'SAVE_INIT';
+export const SAVE_SUCCESS = 'SAVE_SUCCESS';
+export const SAVE_ERROR = 'SAVE_ERROR';